refactor(schedules/utils.py,-schedules/views.py): 스케줄 삭제, 완료 API 리팩토링
사용자 검증 로직 데코레이터로 공통화
