+++ schedules/models.py
@@ -1,4 +1,3 @@
-import stat
 from datetime import timedelta
 
 from django.db import models
